+++ Encountify/ViewModels/BaseViewModel.cs
@@ -1,18 +1,14 @@
-﻿using Encountify.Models;
-using Encountify.Services;
+﻿using Encountify.Services;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Runtime.CompilerServices;
-using System.Threading.Tasks;
 using Xamarin.Forms;
 
 namespace Encountify.ViewModels
 {
     public class BaseViewModel : INotifyPropertyChanged
     {
-        //public IDataStore<Location> DataStore => DependencyService.Get<IDataStore<Location>>();
         public static ILocation LocationData = null;
 
         public BaseViewModel()
@@ -20,7 +16,6 @@ namespace Encountify.ViewModels
             if (LocationData == null)
             {
                 LocationData = DependencyService.Get<ILocation>();
-                LoadLocationDummyData();
             }
         }
 
@@ -49,33 +44,6 @@ namespace Encountify.ViewModels
             onChanged?.Invoke();
             OnPropertyChanged(propertyName);
             return true;
-        }
-
-        public List<Location> DummyLocationList()
-        {
-            LocationData.DeleteAllAsync();
-            return new List<Location>()
-            {
-                new Location("Vilniaus katedra", latitude : 54.685849042698216, longitude :25.287750880122083, category: (int) Category.Cathedral),
-                new Location("Gedimino bokštas", latitude : 54.68667445192699, longitude : 25.29056883194689, category : (int) Category.Castle ),
-                new Location("Vilniaus Šv. Onos bažnyčia", latitude : 54.68378573230062, longitude : 25.292650881640785, category : (int) Category.Church ),
-                new Location("Trys kryžiai", latitude : 54.68740766559662, longitude : 25.29771489238211, category: (int) Category.Monument ),
-                new Location("Gedimino prospektas", latitude : 54.68644019450281, longitude : 25.285441103636185, category : (int) Category.Street),
-                new Location("Lietuvos Respublikos Prezidento kanceliarija", latitude : 54.68383535000863, longitude : 25.286685648648888, category : (int)Category.None),
-                new Location("Sereikiškių parko Bernardinų sodas", latitude : 54.68413305498285, longitude : 25.29522580235671, category : (int) Category.Park ),
-                new Location("Lietuvos nacionalinis dailės muziejus", latitude : 54.68130476957157, longitude : 25.289818468853266, category : (int) Category.Museum ),
-                new Location("Užupio angelas", latitude : 54.68035, longitude : 25.29515, category : (int) Category.Monument),
-                new Location("Pavadinimas", latitude : 54.68, longitude : 25.295, category : (int) Category.SportFacility)
-            };
-        }
-
-        public void LoadLocationDummyData()
-        {
-            var locationList = DummyLocationList();
-            foreach (Location location in locationList)
-            {
-                LocationData.AddAsync(location);
-            }
         }
 
         public event PropertyChangedEventHandler PropertyChanged;
